moving over last bits of ignition
